role & permission

+++ resources/views/themes/original/user/items/index.blade.php
@@ -44,19 +44,24 @@
                 </form>
             </div>
 
-            <div class="d-flex justify-content-end mb-4">
-                <a href="javascript:void(0)" class="btn btn-custom text-white addNewItem"> <i
-                        class="fa fa-plus-circle"></i> @lang('Add Item')</a>
-            </div>
+            @if(adminAccessRoute(array_merge(config('permissionList.Manage_Items.Items.permission.add'))))
+                <div class="d-flex justify-content-end mb-4">
+                    <a href="javascript:void(0)" class="btn btn-custom text-white addNewItem"> <i
+                            class="fa fa-plus-circle"></i> @lang('Add Item')</a>
+                </div>
+            @endif
 
             <div class="table-parent table-responsive me-2 ms-2 mt-4">
                 <table class="table table-striped">
                     <thead>
+
                     <tr>
                         <th scope="col">@lang('SL')</th>
                         <th scope="col">@lang('Item')</th>
                         <th scope="col">@lang('Unit')</th>
-                        <th scope="col">@lang('Action')</th>
+                        @if(adminAccessRoute(array_merge(config('permissionList.Manage_Items.Items.permission.edit'), config('permissionList.Manage_Items.Items.permission.delete'))))
+                            <th scope="col">@lang('Action')</th>
+                        @endif
                     </tr>
 
                     </thead>
@@ -68,7 +73,8 @@
                             <td data-label="@lang('Image')">
                                 <div class="d-flex gap-2">
                                     <div class="logo-brand">
-                                        <img src="{{ getFile(config('location.itemImage.path').$itemList->image) }}" alt="">
+                                        <img src="{{ getFile(config('location.itemImage.path').$itemList->image) }}"
+                                             alt="">
                                     </div>
                                     <div class="product-summary">
                                         <p class="font-weight-bold mt-3">{{ $itemList->name }}</p>
@@ -78,36 +84,43 @@
 
                             <td data-label="@lang('Unit')">{{ $itemList->unit }}</td>
 
-                            <td data-label="Action">
-                                <div class="sidebar-dropdown-items">
-                                    <button
-                                        type="button"
-                                        class="dropdown-toggle"
-                                        data-bs-toggle="dropdown"
-                                        aria-expanded="false"
-                                    >
-                                        <i class="fal fa-cog"></i>
-                                    </button>
+                            @if(adminAccessRoute(array_merge(config('permissionList.Manage_Items.Items.permission.edit'), config('permissionList.Manage_Items.Items.permission.delete'))))
+                                <td data-label="Action">
+                                    <div class="sidebar-dropdown-items">
+                                        <button
+                                            type="button"
+                                            class="dropdown-toggle"
+                                            data-bs-toggle="dropdown"
+                                            aria-expanded="false"
+                                        >
+                                            <i class="fal fa-cog"></i>
+                                        </button>
 
-                                    <ul class="dropdown-menu dropdown-menu-end">
-                                        <li>
-                                            <a class="dropdown-item btn editItem"
-                                               data-route="{{route('user.updateItem', $itemList->id)}}"
-                                               data-property="{{ $itemList }}">
-                                                <i class="fas fa-edit"></i> @lang('Edit')
-                                            </a>
-                                        </li>
+                                        <ul class="dropdown-menu dropdown-menu-end">
+                                            @if(adminAccessRoute(array_merge(config('permissionList.Manage_Items.Items.permission.edit'))))
+                                                <li>
+                                                    <a class="dropdown-item btn editItem"
+                                                       data-route="{{route('user.updateItem', $itemList->id)}}"
+                                                       data-property="{{ $itemList }}">
+                                                        <i class="fas fa-edit"></i> @lang('Edit')
+                                                    </a>
+                                                </li>
+                                            @endif
 
-                                        <li>
-                                            <a class="dropdown-item btn deleteItem"
-                                               data-route="{{route('user.deleteItem', $itemList->id)}}"
-                                               data-property="{{ $itemList }}">
-                                                <i class="fas fa-trash-alt"></i> @lang('Delete')
-                                            </a>
-                                        </li>
-                                    </ul>
-                                </div>
-                            </td>
+                                            @if(adminAccessRoute(array_merge(config('permissionList.Manage_Items.Items.permission.delete'))))
+                                                <li>
+                                                    <a class="dropdown-item btn deleteItem"
+                                                       data-route="{{route('user.deleteItem', $itemList->id)}}"
+
+                                                       data-property="{{ $itemList }}">
+                                                        <i class="fas fa-trash-alt"></i> @lang('Delete')
+                                                    </a>
+                                                </li>
+                                            @endif
+                                        </ul>
+                                    </div>
+                                </td>
+                            @endif
                         </tr>
                     @empty
                         <tr class="text-center">
